requirements.txt
BeautifulSoup==3.2.1
bs4==0.0.1
StringGenerator==0.1.9
pytest==3.6.1
pytest-cache==1.0
pytest-html==1.16.1
pytest-random==0.2
pytest-remove-stale-bytecode==3.0
pytest-timeout==1.2.1
pytest-xdist==1.22.2
pytest-instafail==0.3.0
pytest-repeater==0.1.1
pytest-rerunfailures==4.0
pytest-logger==0.3.0
uiautomator==0.3.2
yapf==0.9.0
pycmd==1.2
pyserial==3.4
pep8==1.7.0
pyflakes==1.6.0
allure-pytest==2.5.0
